+++ modules/importer/data-importer.js
@@ -209,7 +209,7 @@ export default class DataImporter extends FormApplication {
 
       await Promise.all(promises);
       if (isSpecialization) {
-        await this._handleSpecializations(zip);
+        await OggDude.Import.Specializations(zip);
       }
       if (isVehicle) {
         await this._handleVehicles(zip);
@@ -451,159 +451,6 @@ export default class DataImporter extends FormApplication {
       });
     }
     this._importLogger(`Completed Force Power Import`);
-  }
-
-  async _handleSpecializations(zip) {
-    this._importLogger(`Starting Specialization Import`);
-
-    const specializationFiles = Object.values(zip.files).filter((file) => {
-      return !file.dir && file.name.split(".").pop() === "xml" && file.name.includes("/Specializations/");
-    });
-
-    let totalCount = specializationFiles.length;
-    let currentCount = 0;
-
-    if (specializationFiles.length > 0) {
-      $(".import-progress.specializations").toggleClass("import-hidden");
-      let pack = await this._getCompendiumPack("Item", `oggdude.Specializations`);
-
-      await this.asyncForEach(specializationFiles, async (file) => {
-        try {
-          const data = await zip.file(file.name).async("text");
-          const xmlDoc = ImportHelpers.stringToXml(data);
-          const specData = JXON.xmlToJs(xmlDoc);
-
-          let specialization = {
-            name: specData.Specialization.Name,
-            type: "specialization",
-            flags: {
-              ffgimportid: specData.Specialization.Key,
-            },
-            data: {
-              attributes: {},
-              description: specData.Specialization.Description,
-              talents: {},
-              careerskills: {},
-              isReadOnly: true,
-            },
-          };
-
-          specialization.data.description += ImportHelpers.getSources(specData?.Specialization?.Sources ?? specData?.Specialization?.Source);
-          this._importLogger(`Start importing Specialization ${specialization.name}`);
-
-          // assign career skills
-          try {
-            specData.Specialization.CareerSkills.Key.forEach((skillKey) => {
-              let skill = CONFIG.temporary.skills[skillKey];
-
-              if (skill) {
-                // add career skill
-                const careerKey = Object.keys(specialization.data.attributes).length + 1;
-                specialization.data.attributes[`attr${careerKey}`] = {
-                  mod: skill,
-                  modtype: "Career Skill",
-                  value: true,
-                };
-
-                // most specialization give players choice were to put points, create modifier but put value of 0
-                const skillKey = Object.keys(specialization.data.attributes).length + 1;
-                specialization.data.attributes[`attr${skillKey}`] = {
-                  mod: skill,
-                  modtype: "Skill Rank",
-                  value: "0",
-                };
-              }
-            });
-          } catch (err) {
-            // skipping career skills
-          }
-
-          if (specData.Specialization?.Attributes?.ForceRating) {
-            specialization.data.attributes[`attrForceRating`] = {
-              mod: "ForcePool",
-              modtype: "Stat",
-              value: 1,
-            };
-          }
-
-          for (let i = 0; i < specData.Specialization.TalentRows.TalentRow.length; i += 1) {
-            const row = specData.Specialization.TalentRows.TalentRow[i];
-
-            await this.asyncForEach(row.Talents.Key, async (keyName, index) => {
-              let rowTalent = {};
-
-              let talentItem = ImportHelpers.findEntityByImportId("items", keyName);
-              if (!talentItem) {
-                talentItem = await ImportHelpers.findCompendiumEntityByImportId("Item", keyName);
-              }
-
-              if (talentItem) {
-                rowTalent.name = talentItem.data.name;
-                rowTalent.description = talentItem.data.data.description;
-                rowTalent.activation = talentItem.data.data.activation.value;
-                rowTalent.activationLabel = talentItem.data.data.activation.label;
-                rowTalent.isForceTalent = talentItem.data.data.isForceTalent === "true" ? true : false;
-                rowTalent.isConflictTalent = talentItem.data.data.isConflictTalent === "true" ? true : false;
-                rowTalent.isRanked = talentItem.data.data.ranks.ranked === "true" ? true : false;
-                rowTalent.size = "single";
-                rowTalent.canLinkTop = true;
-                rowTalent.canLinkRight = true;
-                rowTalent.itemId = talentItem.data._id;
-                rowTalent.attributes = talentItem.data.data.attributes;
-
-                if (row.Directions.Direction[index].Up && row.Directions.Direction[index].Up === "true") {
-                  rowTalent["links-top-1"] = true;
-                }
-
-                if (row.Directions.Direction[index].Right && row.Directions.Direction[index].Right === "true") {
-                  rowTalent["links-right"] = true;
-                }
-
-                if (talentItem.compendium) {
-                  rowTalent.pack = `${talentItem.compendium.metadata.package}.${talentItem.compendium.metadata.name}`;
-                }
-
-                const talentKey = `talent${i * 4 + index}`;
-                specialization.data.talents[talentKey] = rowTalent;
-              }
-            });
-          }
-
-          // does an image exist?
-          let imgPath = await ImportHelpers.getImageFilename(zip, "Specialization", "", specialization.flags.ffgimportid);
-          if (imgPath) {
-            specialization.img = await ImportHelpers.importImage(imgPath.name, zip, pack);
-          }
-
-          let compendiumItem;
-          await pack.getIndex();
-          let entry = pack.index.find((e) => e.name === specialization.name);
-          if (!entry) {
-            CONFIG.logger.debug(`Importing Specialization - Item`);
-            compendiumItem = new Item(specialization, { temporary: true });
-            this._importLogger(`New Specialization ${specialization.name} : ${JSON.stringify(compendiumItem)}`);
-            pack.importEntity(compendiumItem);
-          } else {
-            CONFIG.logger.debug(`Updating Specialization - Item`);
-            //let updateData = ImportHelpers.buildUpdateData(specialization);
-            let updateData = specialization;
-            updateData["_id"] = entry._id;
-            this._importLogger(`Updating Specialization ${specialization.name} : ${JSON.stringify(updateData)}`);
-            pack.updateEntity(updateData);
-          }
-          currentCount += 1;
-
-          $(".specializations .import-progress-bar")
-            .width(`${Math.trunc((currentCount / totalCount) * 100)}%`)
-            .html(`<span>${Math.trunc((currentCount / totalCount) * 100)}%</span>`);
-          this._importLogger(`End importing Specialization ${specialization.name}`);
-        } catch (err) {
-          CONFIG.logger.error(`Error importing record : `, err);
-        }
-      });
-    }
-
-    this._importLogger(`Completed Specialization Import`);
   }
 
   async _handleVehicles(zip) {
