+++ modules/importer/oggdude/importers/talents.js
@@ -50,7 +50,7 @@ export default class Talents {
 
             data.data.description += ImportHelpers.getSources(item?.Sources ?? item?.Source);
             data.data.attributes = mergeObject(data.data.attributes, ImportHelpers.processStatMod(item?.Attributes));
-            data.data.attributes = mergeObject(data.data.attributes, ImportHelpers.processTalentCareerSkills(item?.ChooseCareerSkills?.NewSkills));
+            data.data.attributes = mergeObject(data.data.attributes, ImportHelpers.processCareerSkills(item?.ChooseCareerSkills?.NewSkills));
             if (item?.DieModifiers) {
               const dieModifiers = await ImportHelpers.processDieMod(item.DieModifiers);
               data.data.attributes = mergeObject(data.data.attributes, dieModifiers.attributes);
